requirements.txt
openai
tiktoken
transformers
torch
requests
python-dotenv
